Move scaling dataset generation and file IO from synthetic_classification_data to scaling_dataset
