+++ specialcouscous/scaling_dataset.py
@@ -1,11 +1,228 @@
+from __future__ import annotations
+
 import argparse
+import logging
 import os
 import pathlib
 from typing import Any
 
-from specialcouscous.synthetic_classification_data import write_scaling_dataset_to_hdf5, generate_scaling_dataset, \
-    read_scaling_dataset_from_hdf5
+import h5py
+import numpy as np
+from sklearn.utils import check_random_state
+
+from specialcouscous.synthetic_classification_data import SyntheticDataset, DatasetPartition
 from specialcouscous.utils import parse_arguments
+
+log = logging.getLogger(__name__)  # Get logger instance.
+
+
+def generate_scaling_dataset(
+    n_samples: int,
+    n_features: int,
+    n_classes: int,
+    n_ranks: int,
+    random_state: int | np.random.RandomState,
+    test_size: float,
+    make_classification_kwargs: dict[str, Any] | None = None,
+    sampling: bool = False,
+    stratified_train_test: bool = False,
+    rank: int | None = None,
+) -> tuple[SyntheticDataset, dict[int, SyntheticDataset] | SyntheticDataset, SyntheticDataset]:
+    """
+    Generate a dataset to be scaled with the number of nodes. Generates a single global dataset and splits it into
+    n_ranks slices. Returns the global train and test set and either the local train set of the given rank, or a dict of
+    all local train sets (mapping rank -> local train set) when rank is None.
+
+    Parameters
+    ----------
+    n_samples : int
+        The number of samples in the dataset.
+    n_features : int
+        The number of features in the dataset.
+    n_classes : int
+        The number classes in the dataset.
+    n_ranks : int
+        The total/maximum number of ranks to generate the dataset for.
+    random_state : int | np.random.RandomState
+        The random state, used for dataset generation, partition, and distribution.
+    test_size : float
+        Relative size of the test set.
+    make_classification_kwargs : dict[str, Any], optional
+        Additional keyword arguments to ``sklearn.datasets.make_classification``.
+    sampling : bool
+        Whether to partition the dataset using deterministic element counts and shuffling or random sampling.
+        See ``SyntheticDataset.get_local_subset`` for more details.
+    stratified_train_test : bool
+        Whether to stratify the train-test split with the class labels.
+    rank : int | None
+        When given, returns the local train set for this rank. Otherwise (i.e. for None), a dict of all local train sets
+        is returned.
+
+    Returns
+    -------
+    SyntheticDataset
+        The global dataset.
+    dict[int, SyntheticDataset] | SyntheticDataset
+        Either the local train subset for this rank (if given) or a dict of all local train sets by rank.
+    SyntheticDataset
+        The global = local test set.
+
+    """
+    random_state = check_random_state(random_state)
+
+    log.debug(f"The random state is:\n{random_state.get_state(legacy=True)}")
+    log.debug(f"Classification kwargs: {make_classification_kwargs}")
+
+    # Step 1: generate balanced global dataset for up to n_ranks nodes
+    global_dataset = SyntheticDataset.generate(
+        n_samples=n_samples,
+        n_features=n_features,
+        n_classes=n_classes,
+        random_state=random_state,
+        make_classification_kwargs=make_classification_kwargs,
+    )
+
+    # Step 2: split into global train and test set
+    # TODO: in this case, the size of the test set scales with n_ranks -> do we want this? if not, what else?
+    log.debug(f"Generate global train-test split: {test_size=}, {stratified_train_test=}.")
+    global_train_set, global_test_set = global_dataset.train_test_split(
+        test_size=test_size,
+        stratify=stratified_train_test,
+        random_state=random_state,
+    )
+    log.debug(f"Shape of global train set {global_train_set.x.shape}, test set {global_test_set.x.shape}")
+
+    # Step 3: partition the global train set into n_ranks local train sets (balanced partition)
+    partition = DatasetPartition(global_train_set.y)
+    assigned_ranks = partition.balanced_partition(n_ranks, random_state, sampling)
+    assigned_indices = partition.assigned_indices_by_rank(assigned_ranks)
+    training_slices = {
+        rank: SyntheticDataset(global_train_set.x[indices], global_train_set.y[indices], None, n_classes)
+        for rank, indices in assigned_indices.items()
+    }
+
+    log.debug(f"Shape of local train set 0 {training_slices[0].x.shape}")
+
+    if rank is not None:
+        log.debug(f"Returning local train set for rank {rank}")
+        return global_train_set, training_slices[rank], global_test_set
+    else:
+        log.debug(f"Returning dict of all local train sets")
+        return global_train_set, training_slices, global_test_set
+
+
+def write_scaling_dataset_to_hdf5(
+    global_train_set: SyntheticDataset,
+    local_train_sets: dict[int, SyntheticDataset],
+    global_test_set: SyntheticDataset,
+    additional_global_attrs: dict[str, Any],
+    file_path: os.PathLike,
+    override: bool = False,
+) -> None:
+    """
+    Write a scaling dataset as HDF5 file to the given path. The HDF5 file has the following structure:
+    Root Attributes:
+    - '/n_classes': the number of classes in the dataset
+    - '/n_ranks': the number of ranks into which the dataset has been partitioned
+    - '/n_samples_global_train': the number of samples in the global training set
+    - **additional_global_attrs: additional_global_attrs are written directly to the root attributes
+      can for example be used to store the config used to generate this dataset
+
+    Groups:
+    - '/test_set': for the global test_set
+    - '/local_train_sets/rank_{rank}': the local train set for each rank
+
+    Group Attributes:
+    - 'n_samples': samples in this subset
+    - 'label': label of the subset (= group name)
+    - 'rank': assigned rank as int (only for local train sets)
+
+
+    Parameters
+    ----------
+    global_train_set : SyntheticDataset
+        The global train set (used only to extract root attributes)
+    local_train_sets : dict[int, SyntheticDataset]
+        pass
+    global_test_set : SyntheticDataset
+        pass
+    additional_global_attrs : dict[str, Any]
+        pass
+    file_path : os.PathLike
+        The file path of the HDF5 file to write.
+    override : bool
+        If true, will override any existing files at file_path. If false, will raise a FileExistsError
+        should file_path already exist.
+    """
+    file_path = pathlib.Path(file_path)
+    if not override and file_path.exists():
+        raise FileExistsError(f"File {file_path} exists and override is set to {file_path}.")
+    file = h5py.File(file_path, "w")
+
+    file.attrs["n_classes"] = global_train_set.n_classes
+    file.attrs["n_ranks"] = len(local_train_sets)
+    file.attrs["n_samples_global_train"] = global_train_set.n_samples
+    for key, value in additional_global_attrs.items():
+        file.attrs[key] = value
+
+    def write_subset_to_group(group_name, dataset, **attrs):
+        file[f'{group_name}/x'] = dataset.x
+        file[f'{group_name}/y'] = dataset.y
+        file[group_name].attrs["n_samples"] = dataset.n_samples
+        for key, value in attrs.items():
+            file[group_name].attrs[key] = value
+
+    # write global test set to HDF5
+    write_subset_to_group("test_set", global_test_set, label="global_test_set")
+
+    # write local train sets to HDF5
+    for rank, local_train_set in local_train_sets.items():
+        group_name = f"local_train_sets/rank_{rank}"
+        write_subset_to_group(group_name, local_train_set, label=group_name, rank=rank)
+
+
+def read_scaling_dataset_from_hdf5(
+    file_path: os.PathLike,
+    rank: int | None = None,
+) -> tuple[dict[int, SyntheticDataset] | SyntheticDataset, SyntheticDataset, dict[str, Any]]:
+    """
+    Read a scaling dataset (local train sets and global test set) from the given HDF5 file.
+
+    Parameters
+    ----------
+    file_path : os.PathLike
+        Path to the HDF5 file to read.
+    rank : int | None
+        Optional rank. When given, only the local train set for the specified rank is retrieved. Otherwise, all local
+        train sets are retrieved.
+
+    Returns
+    -------
+    dict[int, SyntheticDataset] | SyntheticDataset
+        A dict of all local train sets by rank or just the local train set for the given rank.
+    SyntheticDataset
+        The global = local test set.
+    dict[str, Any]
+        Any additional information on the dataset stored as root attribute in the HDF5 file.
+    """
+    file = h5py.File(file_path, "r")
+    n_classes = file.attrs["n_classes"]
+    root_attrs = dict(file.attrs)
+
+    def dataset_from_group(group):
+        return SyntheticDataset(
+            group["x"], group["y"], n_samples=int(group.attrs["n_samples"]), n_classes=int(n_classes)
+        )
+
+    global_test_set = dataset_from_group(file['test_set'])
+
+    if rank is None:  # dict of all local train sets
+        local_train_set = {group.attrs['rank']: dataset_from_group(group)
+                           for name, group in file['local_train_sets'].items()}
+    else:  # just the local train set for the given rank
+        local_train_set = dataset_from_group(file[f'local_train_sets/rank_{rank}'])
+
+    return local_train_set, global_test_set, root_attrs
 
 
 def dataset_path(
